fix: remove storage service abstraction (#530)
the goal of the storage service was to wrap the old minio storage
service with 2 methods for creating and getting presigned urls. it did
this by accessing the inner minio client of that service.

i don't think we need this, we can instead have the minio storage
and new minio storage services implement the same functionality and
have that interface encoded as a new ABC.

so this commit removes the storage service, creates a new ABC,
inherits that ABC in the minio storage services, then has the archive
service to use the minio storage services directly