Merge the `Report._files/_chunks` (#553)
* Merge the `Report._files/_chunks`

Instead of maintaining two separate fields, this merges the previous `_files` which was a dict from filename to `totals`, and the `_chunks` which was either a fully parsed `ReportFile`, or just a shallow `list` of lines.

Maintaining only a single structure for the list of files makes a ton of internals of the `Report` simpler.

It also solves the problem of `file_totals` getting out of sync with the `ReportFile.totals`.

* make parsing/serializing more on-demand

* Move parsing of the `details` to on-demand

* review