Turn `labelanalysis` and `staticanalysis` views into noops (#60)
As part of deprecating ATS, this turns all the views into noops that return empty objects.

Checking in with the requests that the CLI sends, returning an empty `external_id` and `absent_labels` for the labelanalysis will result in the CLI falling back to using the clientside requested labels.

Similarly, returning an empty list of `filepaths` means that the CLI will not do any further uploads.