Add `mock_storage` to a bunch more places (#128)
Running tests with a different (invalid) `config/codecov.yaml` revealed a bunch of tests that are initializing (but not necessarily using) the `ArchiveService`.

I have added the `mock_storage` fixture to those tests, to make sure any test that touches storage (except tests for storage itself) are going through the mock.