Defer merging line coverage to the serialization step (#146)
* Defer merging line coverage to the serialization step

When merging multiple uploads into one, it might be better to construct the per-line coverage from each session at the very end, instead of doing this on each merge.

This should help in particular when multiple uploads cover the same line, and we will go from calling `get_coverage_from_sessions` `O(N)` times for each upload/session to `O(1)`, making the whole operation `O(N)` instead of `O(N^2)` when we have `N` sessions to merge.

As a drive-by change, this removes the `messages` from the lines, as those are completely unused.

* implement this differently to avoid circular imports :-(

* fix

* add a better docstring