fix: remove auto-create indexes (#183)
* fix: remove auto-create indexes

Timescale auto-creates indexes for each group by column paired with
the time bucket column of a continuous aggregate, by default.

We don't want these indexes anymore because:
- they aren't used
- they're causing issues when users attempt to insert testruns with
  long names

the second condition made it so the job to recompute the CAgg from the
base hypertable was failing:

because Postgres has a limitation on the size of entries in a BTree
index, when we would try to add some row with a large name, it would
fail to add an entry into the auto-created (name, timestamp_bin) index
because the entry would exceed the limit imposed by Postgres.

The rows would be inserted into the hypertable with no problem because
the hypertable didn't have an index on the name column.

This same problem applied to the other group by columns: testsuite,
classname, and branch.

In the future, we may want to put an index on the base hypertable so
the group by computation can use an index, and in that case we will
want to change the group by condition to using the test_id, instead of
the testsuite, classname and name.

Originally the test_id was meant to be a column that could be used to
relate the testruns to objects outside of Timescale, like Flake
objects in the test_analytics app, but it would fit the purpose of an
index for the group by as well, since it's just a compressed
representation of the (testsuite, classname, name) tuple.

* chore: update TA timescale indexes

i'm removing some unused indexes on the ta_timeseries_testrun table, and i'm
adding a new one to support the process flakes query for getting testruns

i'm also adding back the indexes to support the queries we do on the summary
tables

* chore: update TA timeseries chunk intervals