POC: endpoint to get evals info (#246)
* POC: endpoint to get evals info

These changes add 2 new endpoints in the REST api (hidden behing a feature flag)
with the objective of being a POC in reporting Eval info saved in Test Analytics.

Probably not what anyone else had in mind, but a POC mind you. I didn't want to deal
with building a frontend. Also the current way _we_ use evals is not in the CI, so I
didn't think it made much sense to mess around the PR comment either.

Anyway the endpoints are:
* summary - aggregates eval data (that is saved per test-run).
* compare - aggregate per commit and compare.

* fix imports because tests exploded in CI

* fix graphQL tests

Apparently there was a custom field lookup in the Database router that was being used in the api.
I moved to the shared one because the api one didn't know about the TA_TIMESERIES database
the shared one didn't know about the special field

I think it's OK to move to just 1, and that should probably be the shared one?
Hopefully. At least tests pass now :E

* typing, docstrings and readability

* renames and logic change

Now all the items are part of the aggregation calculation.
And we export "sum", not only "avg"