feat: fetch TA GQL data directly from Timescale (#311)
* feat: implement direct from Timescale TA aggregates

the previous implementation of calculating TA aggregates relied on a polars
file stored in GCS that contained all the testrun data for a branch in a repo
grouped by test and day. We would aggregate that data further according to the
user's desired time interval and group by test only. We would also calculate
aggregates on all of the data grouped only by repo.

grouped by repo, test, day -> filter by interval -> group by repo, test -> table
                           -> filter by interval -> group by repo -> aggregates

generating this cache file meant that we'd have to scan all rows in Timescale
related to the (branch, repo) combination in the relevant table. These queries
take long and the bottleneck is in reading all the rows from disk, especially if
the data is cold.

What if we could avoid these expensive queries that scan all these rows?

We could fetch only the page of data we need from Timescale to populate the
results in the table, and we could create other continuous aggregates to make it
quick to compute the aggregates we display at the top of the page in the UI.

This commit makes the first step towards the above, by adding code that directly
fetches the relevant GQL fields from Timescale.

The next steps will be to optimize the calculation of the top of the page
aggregates by creating a CAgg that caches the incremental results of that
calculation by day.

I expect that this code might make Timescale performance worse as we'll still
be reading all the rows from the DB to calculate the top of the page aggregates,
but we'll solve this by creating a new CAgg that rolls up the data by day for
the branch and repo, allowing us to only have to read 60 rows per repo in order
to calculate the top of the page aggregates.

* test: fix tests

* fix: be more explicit abt avoiding div by zero

* fix: address feedback