feat: use direct from Timescale TA GQL aggregates behind feature flag (#317)
* feat: use direct from Timescale TA GQL aggregates behind feature flag

See https://github.com/codecov/umbrella/pull/311 for more context

Adding code here to use the new code that fetches TA data directly from
timescale rather than the cache files in GCS. I'm putting it behind the feature
flag because we want to start by rolling this out to ourselves only. We still
want newer users to use Timescale through the cache files though so that's why
I made the changes to `use_new_impl`.

There's also some changes with regards to
the start date and end date, this is for our own sake, since there was some
confusion about what range of days we're getting data for. This is a product
related decision, not a technical decision and has to do with the fact that the
"last day" interval should really mean the last 24 hours, but we don't support
that granularity of data yet.

* test: fix tests