feat(ta-gql): handle receiving all branches filter (#105)
* feat: use new repo wide CAs in GQL API

see:
https://github.com/codecov/umbrella/pull/311
https://github.com/codecov/umbrella/pull/317
https://github.com/codecov/umbrella/pull/319

we're going to use the repo wide CAs to populate some of the data here. We can't
use them for all of the data we want to return in the API so we'll have to fall
back to the testrun branch summaries for those ones: flake_count and
unique_test_count which is subsequently used in the calculation of slow tests
duration.

* ref: TA GQL timescale slowest tests duration implementation

we're replacing the raw SQL query with a django ORM query because it's more
maintainable

* feat: allow filtering TA aggregates by branch

I'm choosing to change the default behaviour when the branch filter is not
specified, even though this will mean that the frontend will inconsistently show
results for all branches on any branch because this code is only deployed to
ourselves for now.

The code change is adding a branch argument to the TestResultsAggregates and
FlakeAggregates GQL types and fetching the data from the correct continuous
aggregates depending on the value of the argument.